### 2019-04-22 #1398 @whedon
[PRE REVIEW]: anesthetic: nested sampling visualisation
**Submitting author:** @williamjameshandley (<a href="http://orcid.org/0000-0002-5866-0445">Will Handley</a>)
**Repository:** <a href="https://github.com/williamjameshandley/anesthetic" target ="_blank">https://github.com/williamjameshandley/anesthetic</a>
**Version:** 1.0.0
