### 2019-04-22 #1398 @danielskatz
👋  @williamjameshandley  - I see a couple of references that are missing - perhaps the cite in the paper.md doesn't quite match the key in the .bib file?  Feel free to make changes, and enter ```@whedon generate pdf``` here as a new comment to rebuild and check.
