### 2019-04-22 #1398 @xuanxu
:wave: @tacaswell: Would you be willing to review this paper? I think your expertise in matplotlib and Python visualization tools makes you a great fit for this submission.
