### 2019-04-22 97031c6
Updated the citations
