+++ paper/paper.bib
@@ -33,8 +33,7 @@ archivePrefix = {arXiv},
 @article{skilling,
 author = "Skilling, John",
 doi = "10.1214/06-BA127",
-fjournal = "Bayesian Analysis",
-journal = "Bayesian Anal.",
+journal = "Bayesian Analysis.",
 month = "12",
 number = "4",
 pages = "833--859",
@@ -111,11 +110,25 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@unpublished{aeons,
+@article{aeons,
    author = {{Handley}, W.~J. and {Speagle}, J. S.},
     title = "{AEONS: Estimating the approximate end of nested sampling}",
   journal = {In preparation}
 }
+
+@article{mcmc,
+ ISSN = {00063444},
+ URL = {http://www.jstor.org/stable/2334940},
+ author = {W. K. Hastings},
+ journal = {Biometrika},
+ number = {1},
+ pages = {97--109},
+ publisher = {[Oxford University Press, Biometrika Trust]},
+ title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
+ volume = {57},
+ year = {1970}
+}
+
 
 @ARTICLE{dnest,
        author = {{Brewer}, Brendon J. and {Foreman-Mackey}, Daniel},
@@ -351,7 +364,7 @@ keywords = "Empirical characteristic function, ECF, Kernel density estimation, H
 }
 
 @INPROCEEDINGS{binder,
-  author = {Jupyter et al.},
+  author = {{Jupyter et al.}},
   title = "{Binder 2.0 - Reproducible, Interactive, Sharable Environments for Science at Scale.}",
   year = "2018",
   doi = {10.25080/Majora-4af1f417-011},
