### 2019-04-22 02eb45b
further bib updates
