+++ paper/paper.bib
@@ -114,6 +114,8 @@ archivePrefix = "arXiv",
    author = {{Handley}, W.~J. and {Speagle}, J. S.},
     title = "{AEONS: Estimating the approximate end of nested sampling}",
   journal = {In preparation}
+     year = 2019,
+    month = may
 }
 
 @article{mcmc,
