+++ paper/paper.bib
@@ -113,7 +113,7 @@ archivePrefix = "arXiv",
 @article{aeons,
    author = {{Handley}, W.~J. and {Speagle}, J. S.},
     title = "{AEONS: Estimating the approximate end of nested sampling}",
-  journal = {In preparation}
+  journal = {In preparation},
      year = 2019,
     month = may
 }
