+++ paper/paper.bib
@@ -110,12 +110,12 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{aeons,
-   author = {{Handley}, W.~J. and {Speagle}, J. S.},
-    title = "{AEONS: Estimating the approximate end of nested sampling}",
-  journal = {In preparation},
-     year = 2019,
-    month = may
+@ARTICLE{aeons,
+       author = {{Handley}, Will and {Speagle}, Joshua},
+        title = "{AEONS: Estimating the approximate end of nested sampling}",
+      journal = {In preparation},
+         year = 2019,
+        month = may,
 }
 
 @article{mcmc,
