### 2019-04-22 371e1ec
Updated bibliography
