+++ paper/paper.bib
@@ -110,12 +110,11 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{aeons,
+@misc{aeons,
        author = {{Handley}, Will and {Speagle}, Joshua},
-        title = "{AEONS: Estimating the approximate end of nested sampling}",
-      journal = {In preparation},
+        title = "{AEONS: Estimating the approximate end of nested sampling (In preparation)}",
          year = 2019,
-        month = may,
+        month = mar,
 }
 
 @article{mcmc,
