+++ paper/paper.bib
@@ -110,11 +110,11 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{aeons,
+@ARTICLE{aeons,
        author = {{Handley}, Will and {Speagle}, Joshua},
-        title = "{AEONS: Estimating the approximate end of nested sampling (In preparation)}",
+        title = "{AEONS: Estimating the approximate end of nested sampling}",
+      journal = {In preparation},
          year = 2019,
-        month = mar,
 }
 
 @article{mcmc,
