+++ paper/paper.bib
@@ -115,6 +115,7 @@ archivePrefix = "arXiv",
         title = "{AEONS: Estimating the approximate end of nested sampling}",
       journal = {In preparation},
          year = 2019,
+         month = mar,
 }
 
 @article{mcmc,
