+++ paper/paper.bib
@@ -115,7 +115,7 @@ archivePrefix = "arXiv",
         title = "{AEONS: Estimating the approximate end of nested sampling}",
       journal = {In preparation},
          year = 2019,
-         month = mar,
+         month = may,
 }
 
 @article{mcmc,
