### 2019-04-22 #1398 @danielskatz
As the assigned editor, @xuanxu will work with you on the reviewers.

From the first comment on this issue, [this list of people](https://bit.ly/joss-reviewers) works for me.

It looks like the link that is generated in the second comment in this issue ([this list of potential reviewers]([])) is broken, however.

Let's see if we can get the code that generated that fixed, by pointing it out to the author :)
