### 2019-04-22 #1398 @danielskatz
👋 @arfon - see the comment above and the broken link that was generated in the second comment in this issue...  Can you fix the underlying code that generated it?
