### 2019-04-23 #1398 @williamjameshandley
@FaustinCarter may be interested in reviewing this, as he maintains a package with some overlap in functionality [pygtc](https://pygtc.readthedocs.io/en/latest/)
