### 2019-04-26 #1398 @xuanxu
@tacaswell, please head over to #1414 for the actual review
