### 2019-05-05 #1414 @tacaswell
- version: there is a 1.0.0 on pypi and a 1.0.0 tag, however there is also a 1.0.1 on pypi, no corosponding tag and ~15 commits on master past the tag.
(https://github.com/williamjameshandley/anesthetic/compare/1.0.0...master).  It looks like there is 1 bug fix and the rest are do
 - Installation: discovered that `fastkde` did not declare it's setup dependencies correctly, opened an issue on their side (https://bitbucket.org/lbl-cascade/fastkde/issues/7/use-setup_requires-and-install_requires-in)
 - the 1D plots don't have yaxis ticks / label.  Is that expected?  (I am concerned that that is actually a Matplotlib bug...)

I recommend publication provisional on the 1.0.0 vs 1.0.1 issue being addressed and confirmation that there should be no y-labels on some of the plots.
