### 2019-05-06 #1414 @williamjameshandley
Many thanks @tacaswell

> version: there is a 1.0.0 on pypi and a 1.0.0 tag, however there is also a 1.0.1 on pypi, no corosponding tag and ~15 commits on master past the tag.
(williamjameshandley/anesthetic@1.0.0...master). It looks like there is 1 bug fix and the rest are documentation.
Indeed. I spotted a minor error in the anesthetic script, and updated the documentation.

>Installation: discovered that fastkde did not declare it's setup dependencies correctly, opened an issue on their side (https://bitbucket.org/lbl-cascade/fastkde/issues/7/use-setup_requires-and-install_requires-in)
Great, thank you very much for this.

> the 1D plots don't have yaxis ticks / label. Is that expected? (I am concerned that that is actually a Matplotlib bug...)
This is an anesthetic feature rather than a  matplotlib bug, since the 1D posterior plots are normalised to have peaks at the same height (as is the usual convention for these kind of triangle plots)
