### 2019-05-10 #1414 @xuanxu
@williamjameshandley:
- There are some references without DOI, could you add the missing DOIs (see [prev comment](https://github.com/openjournals/joss-reviews/issues/1414#issuecomment-491255323)) to the `paper.bib` file?

- The paper mentions version 1.0.0, I've updated the version for the submission to 1.0.1, and created a PR in your repo to update that.

Once you do these minor changes I'll need you to create a Zenodo archive, and report the DOI here, please.
