### 2019-05-10 cfc933d
Added additional doi's
