+++ paper/paper.bib
@@ -127,6 +127,7 @@ archivePrefix = "arXiv",
  pages = {97--109},
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
+ doi = {10.2307/2334940}
  volume = {57},
  year = {1970}
 }
@@ -147,6 +148,23 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/\#abs/2016arXiv160603757B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@Article{gambit,
+author="{The GAMBIT Scanner Workgroup: }",
+title="Comparison of statistical sampling methods with ScannerBit, the GAMBIT scanning module",
+journal="The European Physical Journal C",
+year="2017",
+month="Nov",
+day="13",
+volume="77",
+number="11",
+pages="761",
+issn="1434-6052",
+doi="10.1140/epjc/s10052-017-5274-y",
+url="https://doi.org/10.1140/epjc/s10052-017-5274-y"
+}
+
+
 
 @ARTICLE{dynesty,
        author = {{Speagle}, Joshua S},
@@ -211,14 +229,15 @@ url={http://joss.theoj.org/papers/10.21105/joss.00965}
     adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@misc{zenodo,
-  author       = {Will Handley},
-  title        = {williamjameshandley/anesthetic 1.0.0},
-  month        = apr,
-  year         = 2019,
-  doi          = {10.5281/zenodo.2648394},
-  url          = {https://doi.org/10.5281/zenodo.2648394}
+@article{zenodo, 
+    title={williamjameshandley/anesthetic 1.0.1},
+    DOI={10.5281/zenodo.2668761},
+    publisher={Zenodo},
+    author={Will Handley},
+    year={2019},
+    month={May}
 }
+
 
 @article{pygtc,
     doi = {10.21105/joss.00046},
@@ -331,24 +350,6 @@ archivePrefix = {arXiv},
       archivePrefix  = "arXiv",
       primaryClass   = "astro-ph",
       SLACcitation   = "%%CITATION = ASTRO-PH/0205436;%%"
-}
-
-@ARTICLE{gambit,
-       author = {{GAMBIT Scanner Workgroup}, The and {:} and {Martinez}, Gregory D. and
-         {McKay}, James and {Farmer}, Ben and {Scott}, Pat and
-         {Roebber}, Elinore and {Putze}, Antje and {Conrad}, Jan},
-        title = "{Comparison of statistical sampling methods with ScannerBit, the GAMBIT scanning module}",
-      journal = {arXiv e-prints},
-     keywords = {High Energy Physics - Phenomenology, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Data Analysis, Statistics and Probability, Statistics - Computation},
-         year = "2017",
-        month = "May",
-          eid = {arXiv:1705.07959},
-        pages = {arXiv:1705.07959},
-archivePrefix = {arXiv},
-       eprint = {1705.07959},
- primaryClass = {hep-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2017arXiv170507959G},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 
