### 2019-05-10 #1414 @williamjameshandley
@xuanxu I've merged in your changes to the paper, and added the DOIs that exist. Unfortunately there aren't any for the DNest 4 or MontePython papers, as they're only arXiv articles
