### 2019-05-10 #1414 @williamjameshandley
> Once you do these minor changes I'll need you to report the new Zenodo archive's DOI here, please.

Version 1.0.1 10.5281/zenodo.2668761
