+++ paper/paper.bib
@@ -127,7 +127,7 @@ archivePrefix = "arXiv",
  pages = {97--109},
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
- doi = {10.2307/2334940}
+ doi = {10.2307/2334940},
  volume = {57},
  year = {1970}
 }
