### 2019-05-12 #1414 @xuanxu
Thanks @williamjameshandley!

There is a typo in the bib file preventing the paper to compile: a missing comma, I've issued a PR in your repo fixing it.

> Unfortunately there aren't any for the DNest 4 or MontePython papers, as they're only arXiv articles

So the DOIs [suggested by whedon](https://github.com/openjournals/joss-reviews/issues/1414#issuecomment-491255323) for them are not the right ones?

> ```
> MISSING DOIs
>
> - https://doi.org/10.18637/jss.v086.i07 may be missing for title: DNest4: Diffusive Nested Sampling in C++ and Python
> - https://doi.org/10.1016/j.dark.2018.100260 may be missing for title: MontePython 3: boosted MCMC sampler and other features
> ```
