### 2019-05-12 #1414 @williamjameshandley
> So the DOIs suggested by whedon for them are not the right ones?

Ah -- that's very clever, I missed the autosuggestions. I've added in the full bibtex associated with those doi links.
