### 2019-05-12 #1414 @xuanxu
@williamjameshandley Do you mind to create a new zenodo archive (and update the reference in the bib file )?
As it is is ok because there's no changes in the code, but I'd love the reference to point to a deposit that includes the correct files for the final joss paper. WDYT?
