### 2019-05-12 #1414 @williamjameshandley
Sure.

What is the correct way to do this? I'm struggling to see how I could have a bib file with the correct doi archived in the corresponding zenodo doi, unless I use the 'all versions' doi 10.5281/zenodo.2614017
