### 2019-05-12 #1414 @xuanxu
You are right. I think it's ok to have the link in the paper using the 'all versions' doi. It will redirect readers to the more recent version.
Report here the DOI of the latest version and I'll use that here in the JOSS submission.
