### 2019-05-12 #1414 @xuanxu
The other option is to create a deposit and then change the bib file to point to it. That way the diference between the final paper and the deposit is minimal.
Whatever option you prefer is OK.
