### 2019-05-12 11ce6b3
Updated version number for completed JOSS paper
