+++ paper/paper.bib
@@ -229,8 +229,8 @@ url={http://joss.theoj.org/papers/10.21105/joss.00965}
 }
 
 @article{zenodo, 
-    title={williamjameshandley/anesthetic 1.0.1},
-    DOI={10.5281/zenodo.2668761},
+    title={williamjameshandley/anesthetic 1.0.2},
+    DOI={10.5281/zenodo.2614017},
     publisher={Zenodo},
     author={Will Handley},
     year={2019},
