### 2019-05-12 #1414 @williamjameshandley
I've gone with the first option, as it works better with GitHub's syncing between releases and zenodo archives. I updated the version number to 1.0.2 in the README, paper and bib file, pushed a new pypi version, and created a new github release. The 'all versions' DOI is:

10.5281/zenodo.2614017

Many thanks @xuanxu
