### 2019-05-12 #1414 @xuanxu
This is ready now for publication.
Pinging EIC for final acceptance: @openjournals/joss-eics
