### 2019-05-12 #1414 @labarba
- Use the common capitalization of: Python, NumPy, SciPy, and Matplotlib (pandas is usually not capitalized).
- Check your [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax). If citing two works at once, separate by semicolon to get them in the same parenthesis group.
- "Currently public available implementations" >> awkward. Maybe "Currently available public implementations"?
- "all which have been incorporated" >> all _of_ which
- on page 2, you're missing a space before citations, so the parenthesis touches the previous word
- "one and two-dimensional" >> add hyphen after "one"
- "A subset of computations from item 1 are provided" >> _is_ provided
- "the effect of adjust dynamically adjusting " >> typo

Just before the Acknowledgements, you have a link to the code repository. We include a link to the repo on the margin decorators (first page), so we ask that you don't add an additional link in the text or references. The Zenodo archive will also be linked on the first page. Can you edit those out?

The notebook is not "hosted" on Binder ... it is hosted on GitHub but Binder can build an image on request. Please reword.
