### 2019-05-12 #1414 @labarba
You should add caps protection on `{mcmc}` here: https://github.com/williamjameshandley/anesthetic/blob/11ce6b382c5a528e60117adba32d71de8c993efa/paper/paper.bib#L324
