### 2019-05-12 #1414 @labarba
This DOI is not resolving:
https://github.com/williamjameshandley/anesthetic/blob/11ce6b382c5a528e60117adba32d71de8c993efa/paper/paper.bib#L330

Try putting _just_ the DOI, i.e., the alphanumeric identifier, not the full URL, into the `doi` field of the `.bib` file.
