### 2019-05-12 #1414 @labarba
Handley, W., & Speagle, J. (2019) — do not cite items that are not publicly available. Either upload a preprint to arXiv, or remove the citation (it is OK to mention in the text that you have some work in progress, but not to cite it).
