### 2019-05-12 #1414 @labarba
We need caps protection on "Carlo" and "Markov" in this title:
https://github.com/williamjameshandley/anesthetic/blob/11ce6b382c5a528e60117adba32d71de8c993efa/paper/paper.bib#L129
