### 2019-05-12 #1414 @labarba
Caps protection on "Bayes"
https://github.com/williamjameshandley/anesthetic/blob/11ce6b382c5a528e60117adba32d71de8c993efa/paper/paper.bib#L41
