+++ paper/paper.md
@@ -24,102 +24,48 @@ bibliography: paper.bib
 ---
 
 # Summary
-``anesthetic`` is a python package for processing nested sampling runs, and
-will be useful for any scientist or statistician who uses nested sampling
-software. ``anesthetic`` unifies many existing tools and techniques in an
-extensible framework which is intuitive for users familiar with the standard
-python packages, namely [numpy](https://www.numpy.org/),
-[scipy](https://www.scipy.org/), [matplotlib](https://matplotlib.org/) and
-[pandas](https://pandas.pydata.org/). It has been extensively used in the
-recent cosmological papers [@tension][@dimensionality].
+``anesthetic`` is a Python package for processing nested sampling runs, and will be useful for any scientist or statistician who uses nested sampling software. ``anesthetic`` unifies many existing tools and techniques in an extensible framework which is intuitive for users familiar with the standard Python packages, namely [NumPy](https://www.numpy.org/), [SciPy](https://www.scipy.org/), [Matplotlib](https://matplotlib.org/) and [pandas](https://pandas.pydata.org/). It has been extensively used in the recent cosmological papers [@tension;@dimensionality].
 
 
 
 # Nested sampling
 
-Nested sampling [@skilling] is an alternative to Markov-Chain-Monte-Carlo
-techniques [@mcmc]. Given some data $D$, for a scientific model $M$ with free
-parameters $\theta$, Bayes theorem states:
+Nested sampling [@skilling] is an alternative to Markov-Chain-Monte-Carlo techniques [@mcmc]. Given some data $D$, for a scientific model $M$ with free parameters $\theta$, Bayes theorem states:
 
 $$ P(\theta|D) = \frac{P(D|\theta) P(\theta)}{P(D)}. $$
 
-Traditional MCMC approaches ignore the Bayesian evidence $P(D)$ and instead
-focus on the problem of generating samples from the posterior $P(\theta|D)$
-using knowledge of the prior $P(\theta)$ and likelihood $P(D|\theta)$. Nested
-sampling reverses this priority, and instead computes the evidence $P(D)$ (the
-critical quantity in Bayesian model comparison [@trotta]), producing posterior
-samples as a by-product. Nested sampling does this by evolving a set of live
-points drawn from the prior under a hard likelihood constraint which steadily
-increases, causing the live points to contract around the peak(s) of the
-likelihood. The history of the live-point evolution can be used to reconstruct
-both the evidence and posterior samples, as well as the density of states and
-consequently the full partition function [@aeons].
+Traditional MCMC approaches ignore the Bayesian evidence $P(D)$ and instead focus on the problem of generating samples from the posterior $P(\theta|D)$ using knowledge of the prior $P(\theta)$ and likelihood $P(D|\theta)$. Nested sampling reverses this priority, and instead computes the evidence $P(D)$ (the critical quantity in Bayesian model comparison [@trotta]), producing posterior samples as a by-product. Nested sampling does this by evolving a set of live points drawn from the prior under a hard likelihood constraint which steadily increases, causing the live points to contract around the peak(s) of the likelihood. The history of the live-point evolution can be used to reconstruct both the evidence and posterior samples, as well as the density of states and consequently the full partition function.
 
-Currently public available implementations of nested sampling include MultiNest
-[@multinest], PolyChord [@polychord0][@polychord1][@dypolychord], DNest
-[@dnest] and dynesty [@dynesty], all which have been incorporated into a
-wide range of cosmological [@cosmomc][@cosmosis][@montepython] and particle
-physics [@gambit] codes.
[…]
