+++ paper/paper.bib
@@ -46,7 +46,7 @@ year = "2006"
 
 @ARTICLE{trotta,
    author = {{Trotta}, R.},
-    title = "{Bayes in the sky: Bayesian inference and model selection in cosmology}",
+    title = "{{Bayes} in the sky: Bayesian inference and model selection in cosmology}",
   journal = {Contemporary Physics},
 archivePrefix = "arXiv",
    eprint = {0803.4089},
@@ -110,14 +110,6 @@ archivePrefix = "arXiv",
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{aeons,
-       author = {{Handley}, Will and {Speagle}, Joshua},
-        title = "{AEONS: Estimating the approximate end of nested sampling}",
-      journal = {In preparation},
-         year = 2019,
-         month = may,
-}
-
 @article{mcmc,
  ISSN = {00063444},
  URL = {http://www.jstor.org/stable/2334940},
@@ -126,7 +118,7 @@ archivePrefix = "arXiv",
  number = {1},
  pages = {97--109},
  publisher = {[Oxford University Press, Biometrika Trust]},
- title = {Monte Carlo Sampling Methods Using Markov Chains and Their Applications},
+ title = {Monte {Carlo} Sampling Methods Using {Markov} Chains and Their Applications},
  doi = {10.2307/2334940},
  volume = {57},
  year = {1970}
@@ -183,7 +175,7 @@ archivePrefix = {arXiv},
 
 
 @article{dypolychord,
-title={dyPolyChord: dynamic nested sampling with PolyChord},
+title={{dyPolyChord}: dynamic nested sampling with {PolyChord}},
 author={Higson, Edward},
 year={2018},
 journal={Journal of Open Source Software},
@@ -226,15 +218,6 @@ url={http://joss.theoj.org/papers/10.21105/joss.00965}
     month = apr,
     adsurl = {http://adsabs.harvard.edu/abs/2018arXiv180406406H},
     adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@article{zenodo, 
-    title={williamjameshandley/anesthetic 1.0.2},
-    DOI={10.5281/zenodo.2614017},
-    publisher={Zenodo},
-    author={Will Handley},
-    year={2019},
-    month={May}
 }
 
 
@@ -321,13 +304,13 @@ archivePrefix = {arXiv},
 }
 
 @article{montepython,
-title = "MontePython 3: Boosted MCMC sampler and other features",
+title = "MontePython 3: Boosted {MCMC} sampler and other features",
 journal = "Physics of the Dark Universe",
 volume = "24",
 pages = "100260",
 year = "2019",
 issn = "2212-6864",
-doi = "https://doi.org/10.1016/j.dark.2018.100260",
+doi = "10.1016/j.dark.2018.100260",
 url = "http://www.sciencedirect.com/science/article/pii/S2212686418302309",
 author = "Thejs Brinckmann and Julien Lesgourgues",
 keywords = "Cosmology, Parameter inference, Numerical tools",
