### 2019-05-12 #1414 @williamjameshandley
Many thanks @labarba, I believe I've implemented all those corrections.
