+++ paper/paper.bib
@@ -340,7 +340,7 @@ volume = "101",
 pages = "148 - 160",
 year = "2016",
 issn = "0167-9473",
-doi = "https://doi.org/10.1016/j.csda.2016.02.014",
+doi = "10.1016/j.csda.2016.02.014",
 url = "http://www.sciencedirect.com/science/article/pii/S0167947316300408",
 author = "Travis A. O’Brien and Karthik Kashinath and Nicholas R. Cavanaugh and William D. Collins and John P. O’Brien",
 keywords = "Empirical characteristic function, ECF, Kernel density estimation, Histogram, Nonuniform FFT, NuFFT, Multidimensional, KDE",
