### 2019-05-12 #1414 @williamjameshandley
> This DOI link is also broken:
> https://github.com/williamjameshandley/anesthetic/blob/11ce6b382c5a528e60117adba32d71de8c993efa/paper/paper.bib#L360

Apologies, I missed this one. Latest push has this fixed as well.
