### 2019-05-12 #1414 @labarba
Please accept my small edits, via PR. Note that the PDF generated above still shows one broken DOI link, but when I open the `.bib` file it looks correct. Unsure what's up.
