+++ paper/paper.bib
@@ -335,7 +335,7 @@ abstract = "MontePython is a parameter inference package for cosmology. We prese
 
 @article{fastkde,
 title = "A fast and objective multidimensional kernel density estimation method: fastKDE",
-journal = "Computational Statistics & Data Analysis",
+journal = {Computational Statistics \& Data Analysis},
 volume = "101",
 pages = "148 - 160",
 year = "2016",
