### 2019-05-12 #1414 @williamjameshandley
> Please accept my small edits, via PR. Note that the PDF generated above still shows one broken DOI link, but when I open the `.bib` file it looks correct. Unsure what's up.

I'm not exactly sure which one you're referring to, but I noticed that there was an errant ``&`` in the fastKDE journal name. I changed it to ``\&`` and it seems to be resolving better, and the DOI no longer has http in front of it.
