### 2019-05-12 #1414 @williamjameshandley
(Edits merged)
