### 2019-05-12 #1414 @labarba
Congratulations, @williamjameshandley, your JOSS paper is published 🚀

Huge thanks to our editor: @xuanxu, and reviewer: @tacaswell — we couldn't do this without you 🙏
