### 2019-05-12 #1414 @williamjameshandley
Many thanks for all your efforts @labarba, @xuanxu and @tacaswell: I remain very impressed by the quality, efficiency and usefulness of the JOSS review process.
